feat(utils): split serialization and improve performance
