+++ snark-utils/src/helpers.rs
@@ -1,64 +1,31 @@
 use crate::errors::{Error, VerificationError};
-use crate::{buffer_size, Deserializer, Serializer, UseCompression};
+use crate::{buffer_size, BatchSerializer, Result, UseCompression};
 use blake2::{digest::generic_array::GenericArray, Blake2b, Digest};
 use crypto::digest::Digest as CryptoDigest;
 use crypto::sha2::Sha256;
 use rand::{rngs::OsRng, thread_rng, Rng, SeedableRng};
 use rand_chacha::ChaChaRng;
 use std::convert::TryInto;
-use std::io::{self, Read, Write};
+use std::io::{self, Write};
 use std::ops::{AddAssign, Mul};
 
 use std::sync::Arc;
 use typenum::consts::U64;
 use zexe_algebra::{
-    AffineCurve, BigInteger, CanonicalSerialize, Field, One, PairingEngine, PrimeField,
-    ProjectiveCurve, UniformRand, Zero,
+    AffineCurve, BigInteger, CanonicalSerialize, ConstantSerializedSize, Field, One, PairingEngine,
+    PrimeField, ProjectiveCurve, UniformRand, Zero,
 };
 
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
-
 use zexe_fft::{cfg_into_iter, cfg_iter, cfg_iter_mut};
-
-/// A convenience result type for returning errors
-pub type Result<T> = std::result::Result<T, Error>;
-
-/// Allocates a buffer to read the bytes from the provided reader
-/// and then reads the element from the buffer
-// TODO: Make this not do 2 allocations
-pub fn read_element<C: AffineCurve, R: Read>(
-    reader: &mut R,
-    compression: UseCompression,
-) -> Result<C> {
-    let mut buf = vec![0; buffer_size::<C>(compression)];
-    reader.read_exact(&mut buf)?;
-    let element = buf.read_element::<C>(compression)?;
-    // reject points at infinity
-    if element.is_zero() {
-        return Err(Error::PointAtInfinity);
-    }
-    Ok(element)
-}
-
-/// Allocates a buffer to serialize the provided element
-/// and then writes that buffer to the provided reader
-pub fn write_element<C: AffineCurve, W: Write>(
-    writer: &mut W,
-    element: &C,
-    compression: UseCompression,
-) -> Result<()> {
-    let mut buf = vec![0; buffer_size::<C>(compression)];
-    buf.write_element(element, compression)?;
-    writer.write_all(&buf)?;
-    Ok(())
-}
 
 pub fn write_elements<C: AffineCurve, W: Write>(
     writer: &mut W,
     elements: &[C],
     compression: UseCompression,
 ) -> Result<()> {
+    // we want to write to the buffer in parallel, so we'll have to allocate here to [u8]
     let mut buf = vec![0; elements.len() * buffer_size::<C>(compression)];
     buf.write_batch(&elements, compression)?;
     writer.write_all(&buf)?;
@@ -432,10 +399,10 @@ pub fn compute_g2_s<E: PairingEngine>(
     let mut h = Blake2b::default();
     h.input(&[personalization]);
     h.input(digest);
-    let size = E::G1Affine::buffer_size();
+    let size = E::G1Affine::constant_serialized_size();
     let mut data = vec![0; 2 * size];
-    g1_s.serialize(&[], &mut data[..size])?;
-    g1_s_x.serialize(&[], &mut data[size..])?;
+    g1_s.serialize(&mut &mut data[..size])?;
+    g1_s_x.serialize(&mut &mut data[size..])?;
     h.input(&data);
     Ok(hash_to_g2::<E>(h.result().as_ref()).into_affine())
 }
