+++ snark-utils/src/lib.rs
@@ -6,6 +6,9 @@
 pub mod errors;
 pub use errors::{Error, InvariantKind, Phase2Error, VerificationError};
 
+/// A convenience result type for returning errors
+pub type Result<T> = std::result::Result<T, Error>;
+
 mod groth16_utils;
 pub use groth16_utils::Groth16Params;
 
@@ -16,7 +19,7 @@ mod helpers;
 pub use helpers::*;
 
 mod io;
-pub use io::{buffer_size, Deserializer, Serializer};
+pub use io::{buffer_size, BatchDeserializer, BatchSerializer, Deserializer, Serializer};
 
 // Re-exports for handling hashes
 pub use blake2::digest::generic_array::GenericArray;
